initial pass of topology extraction
