--- goldstone/glance/models.py
@@ -1,21 +1,0 @@
-"""Glance models."""
-# Copyright 2015 Solinea, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from goldstone.models import TopologyData
-
-
-class ImagesData(TopologyData):
-    """Return data from ES about glance images"""
-    _DOC_TYPE = 'glance_image_list'
-    _INDEX_PREFIX = 'goldstone-'
