--- goldstone/glance/tests.py
@@ -1,42 +1,0 @@
-"""Glance unit tests."""
-# Copyright 2015 Solinea, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from django.test import SimpleTestCase
-
-
-class DataViewTests(SimpleTestCase):
-    """Test the data view."""
-
-    def _evaluate(self, response):
-        """Check the response."""
-        import json
-        from django.http import HttpResponse
-
-        self.assertIsInstance(response, HttpResponse)
-        self.assertIsNotNone(response.content)
-        j = json.loads(response.content)
-        self.assertIsInstance(j, list)
-
-    def test_get_images(self):
-        """GET to /images."""
-        from django.contrib.auth import get_user_model
-        from goldstone.test_utils import create_and_login, \
-            AUTHORIZATION_PAYLOAD
-
-        get_user_model().objects.all().delete()
-        token = create_and_login()
-
-        self._evaluate(
-            self.client.get("/glance/images/",
-                            HTTP_AUTHORIZATION=AUTHORIZATION_PAYLOAD % token))
