--- goldstone/glance/urls.py
@@ -1,22 +1,0 @@
-"""Glance app URLconf."""
-# Copyright 2015 Solinea, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from django.conf.urls import url, patterns
-from .views import ImagesDataView
-
-# Views handled by DjangoRestFramework Views.
-urlpatterns = patterns(
-    '',
-    url(r'^images', ImagesDataView.as_view(), name='glance-images'),
-    )
