--- goldstone/glance/views.py
@@ -1,37 +1,0 @@
-"""Glance views."""
-# Copyright 2015 Solinea, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from goldstone.core.utils import JsonReadOnlyView
-from .models import ImagesData
-
-
-# Our API documentation extracts this docstring, hence the use of markup.
-class ImagesDataView(JsonReadOnlyView):
-    """Return Images data.
-
-    ---
-
-    GET:
-        parameters:
-           - name: region
-             description: The desired region.
-             paramType: query
-           - name: zone
-             description: The desired zone.
-             paramType: query
-
-    """
-
-    model = ImagesData
-    key = 'images'
