--- goldstone/nova/fixtures/nova_initial_data.yaml
@@ -1,35 +1,0 @@
-  - model: core.SavedSearch
-    pk: 21f5c6db-5a2e-41d4-9462-c3cdc03a837b
-    fields:
-      name: nova spawns
-      owner: core
-      index_prefix: goldstone-*
-      doc_type: nova_spawns
-      timestamp_field: '@timestamp'
-      protected: true
-      query: '{
-   "query": {
-      "term": {
-         "event": "finish"
-      }
-   },
-   "aggs": {
-      "per_interval": {
-         "date_histogram": {
-            "field": "@timestamp",
-            "interval": "1d",
-            "min_doc_count": 0
-         },
-         "aggs": {
-            "success": {
-               "terms": {
-                  "field": "success.raw",
-                  "shard_min_doc_count": 0,
-                  "min_doc_count": 0,
-                  "size": 0
-               }
-            }
-         }
-      }
-   }
-}'
